amazon_shopping/requirements.txt
openai-agents>=0.0.7
python-dotenv>=1.0.0
